fix(client): replace Cubism 5.3 SDK with 5-r.2 (no URP dependency)
Cubism 5.3 embeds URP render-pass/blend-mode code in its main renderer
(CubismRenderer.cs references DrawObjectType/Offscreen/RenderGraph), so it
cannot compile without URP - which the Tuanjie 1.2.0 registry lacks.
5-r.2 has zero URP references and matches Unity 2022.3 built-in RP.
- removed 5.3-only files (Rendering/URP, *UsingBlendMode partials)
- imported 5-r.2 source (183 cs files) from Live2D/CubismUnityComponents

+++ zerolan-vtuber/client/Assets/Live2D/Cubism/Samples/OriginalWorkflow/Motion/CubismMotionPreview.cs
@@ -46,7 +46,7 @@ namespace Live2D.Cubism.Samples.OriginalWorkflow.Motion
 
 
 
-        private void PlayIdleAnimation(int index = 0)
+        private void PlayIdleAnimation(float index = 0.0f)
         {
             _motionController.PlayAnimation(Animation, isLoop: false, priority: CubismMotionPriority.PriorityIdle);
         }
